+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/components/Tabs/PolicyTab/PolicyTab.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from 'react';
-import { PolicyEditor as PolicyEditorImpl, PolicySubject } from '@altinn/policy-editor';
+import { PolicyEditor as PolicyEditorImpl } from '@altinn/policy-editor';
 import { useAppPolicyMutation } from 'app-development/hooks/mutations';
 import { useTranslation } from 'react-i18next';
 import { TabHeader } from '../../TabHeader';
@@ -12,13 +12,6 @@ import {
   useResourcePolicySubjectsQuery,
 } from 'app-shared/hooks/queries';
 import { mergeQueryStatuses } from 'app-shared/utils/tanstackQueryUtils';
-
-const policySubjectOrg: PolicySubject = {
-  subjectDescription: '[org]',
-  subjectId: '[org]',
-  subjectSource: '[org]',
-  subjectTitle: '[org]',
-};
 
 export type PolicyTabProps = {
   /**
@@ -57,16 +50,9 @@ export const PolicyTab = ({ org, app }: PolicyTabProps): ReactNode => {
     status: subjectStatus,
     data: subjectData,
     error: subjectError,
-  } = useResourcePolicySubjectsQuery(org, app, true);
+  } = useResourcePolicySubjectsQuery(org, app);
 
   const { mutate: updateAppPolicyMutation } = useAppPolicyMutation(org, app);
-
-  const addOrgIfNotAlreadyIn = (subjects: PolicySubject[]) => {
-    if (!subjects.some((d) => d.subjectId === policySubjectOrg.subjectId)) {
-      return [...subjects, policySubjectOrg];
-    }
-    return subjects;
-  };
 
   const displayContent = () => {
     switch (mergeQueryStatuses(policyStatus, actionStatus, subjectStatus)) {
@@ -87,7 +73,7 @@ export const PolicyTab = ({ org, app }: PolicyTabProps): ReactNode => {
           <PolicyEditorImpl
             policy={policyData}
             actions={actionData}
-            subjects={addOrgIfNotAlreadyIn(subjectData)}
+            subjects={subjectData}
             onSave={updateAppPolicyMutation}
             showAllErrors
             usageType='app'
