+++ frontend/packages/shared/src/hooks/queries/useResourcePolicySubjectsQuery.ts
@@ -4,6 +4,13 @@ import type { PolicySubject } from '@altinn/policy-editor';
 import { QueryKey } from 'app-shared/types/QueryKey';
 import { AxiosError } from 'axios';
 
+const policySubjectOrg: PolicySubject = {
+  subjectDescription: '[org]',
+  subjectId: '[org]',
+  subjectSource: '[org]',
+  subjectTitle: '[org]',
+};
+
 /**
  * Query to get the list of subjects for a policy.
  *
@@ -20,7 +27,17 @@ export const useResourcePolicySubjectsQuery = (
 ): UseQueryResult<PolicySubject[], AxiosError> => {
   const { getPolicySubjects } = useServicesContext();
 
-  return useQuery<PolicySubject[], AxiosError>([QueryKey.ResourcePolicySubjects, org, repo], () =>
-    getPolicySubjects(org, repo),
+  return useQuery<PolicySubject[], AxiosError>(
+    [QueryKey.ResourcePolicySubjects, org, repo],
+    () => getPolicySubjects(org, repo),
+    {
+      onSuccess: (data: PolicySubject[]) => {
+        console.log(data);
+
+        if (addOrgToList && !data.some((d) => d.subjectId === policySubjectOrg.subjectId))
+          data.push(policySubjectOrg);
+        return data;
+      },
+    },
   );
 };
